+++ lib/app.js
@@ -4,7 +4,7 @@ const chalk = require('chalk');
 const Watcher = require('./watcher');
 const { copyDir } = require('./utils/fs-utils');
 const { zipFiles } = require('./utils/zip');
-const ExcelFilePipeline = require('./excel-file-pipeline');
+const ExcelFilePipeline = require('./excel-file-manager');
 const { createMockProcessor, parseExcelIntoMocks } = require('./mock-processings');
 
 class App {
