+++ lib/excel-file-manager.js
@@ -5,7 +5,10 @@ const assert = require('assert');
 const crypto = require('crypto');
 const { readFileAsync } = require('./utils/fs-utils');
 
-class ExcelFilePipeline extends EventEmitter {
+class ExcelFileManager extends EventEmitter {
+
+
+
     /**
      * Create File pipeline
      * @param {string} destRoot destination folder
@@ -85,4 +88,4 @@ class ExcelFilePipeline extends EventEmitter {
     }
 }
 
-module.exports = ExcelFilePipeline;
+module.exports = ExcelFileManager;
