+++ lib/include-file-manager.js
@@ -30,7 +30,7 @@ class IncludeFileManager {
     get fileHashMap() { return this.#fileHashMap }
     get includeDirs() { return [...this.#includeDirs] }
 
-    async processDir() {
+    async includeDir() {
         this.#initResults();
         await this.#copyDir(this.#includeRoot, this.#destDir);
     }
