+++ lib/tests/excel-file-manager.test.js
@@ -34,7 +34,7 @@ jest.mock('fs', () => {
         _getFileContent: () => files,
     };
 });
-const FilePipeline = require('../excel-file-pipeline');
+const FilePipeline = require('../excel-file-manager');
 
 
 test('should process files', async () => {
