+++ lib/tests/include-file-manager.test.js
@@ -14,7 +14,7 @@ test('should copy files and return proper results w/o hash', async () => {
     }, '/includes');
 
     const im = new IncludeFileManager({destDir: '/dest', includeDir: '/includes'});
-    await im.processDir();
+    await im.includeDir();
 
     expect(im.includedFiles).toEqual([
         'INCLUDE1.txt',
@@ -45,7 +45,7 @@ test('should copy files and return proper results with hashing', async () => {
     }, '/includes');
 
     const im = new IncludeFileManager({destDir: '/dest', includeDir: '/includes', withHashing: true});
-    await im.processDir();
+    await im.includeDir();
 
     expect(im.includedFiles).toEqual([
         'INCLUDE1.txt',
@@ -78,7 +78,7 @@ test('should copy one file update hash', async () => {
     }, '/includes');
 
     const im = new IncludeFileManager({destDir: '/dest', includeDir: '/includes', withHashing: true});
-    await im.processDir();
+    await im.includeDir();
 
     vol.writeFileSync('/includes/INCLUDE1.txt', 'X');
     vol.writeFileSync('/includes/INCLUDE2.txt', 'Y');
@@ -111,7 +111,7 @@ test('should copy one file and update dirs', async () => {
     }, '/includes');
 
     const im = new IncludeFileManager({destDir: '/dest', includeDir: '/includes', withHashing: true});
-    await im.processDir();
+    await im.includeDir();
 
     vol.mkdirSync('/includes/sub1');
     vol.writeFileSync('/includes/sub1/include2.txt', '2');
@@ -144,7 +144,7 @@ test('should throw on file with another root', async () => {
     }, '/includes');
 
     const im = new IncludeFileManager({destDir: '/dest', includeDir: '/includes', withHashing: true});
-    await im.processDir();
+    await im.includeDir();
 
     await expect(im.includeOneFile('/another/include3.txt')).rejects.toThrowError('must be relative');
 });
